\name{tempsuitcalc-package}
\alias{tempsuitcalc-package}
\alias{tempsuitcalc}
\docType{package}

\title{
Simulates mosquito cohorts and calculates vec capacity
}
\description{
Simulates mosquito cohorts and calculates vec capacity
}
\details{
\tabular{ll}{
Package: \tab tempsuitcalc\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-09-10\cr
License: \tab GPL (>=2)\cr
}

}
\author{
Oliver Brady <oliver.brady@zoo.ox.ac.uk>
}

\seealso{
\code{\link{cohort.simulate}}
}

